requirements.txt
Django==2.2rc1
djangorestframework==3.9.2
